course update api added

+++ ELearning/src/main/java/com/elearning/dao/CourseDAO.java
@@ -7,7 +7,7 @@ import com.elearning.model.Course;
 public interface CourseDAO {
 
 	Long createCourse(Course course);
-	Long updateCourse(Course course);
+	Integer updateCourse(Course course);
 	List<Course> getAllCourses();
 	Course getCourseById(Long id);
 }
